requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.1
chardet==3.0.4
code-snippet==2.0.0
Django==3.0.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-filter==2.2.0
django-picklefield==3.0.1
djangorestframework==3.11.0
future==0.18.2
idna==2.9
importlib-metadata==1.6.1
kombu==4.6.10
Markdown==3.2.1
mbed-tools-lib==1.2.0
numpy==1.20.1
pandas==1.2.2
pystache==0.5.4
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
redis==3.5.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typing-extensions==3.7.4.3
Unipath==1.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
